# Optimize the cat operation on contiguous tensors