--- a/candle-core/src/backend.rs
+++ b/candle-core/src/backend.rs
@@ -99,6 +99,7 @@ pub trait BackendStorage: Sized {
 
     fn copy_strided_src(&self, _: &mut Self, _: usize, _: &Layout) -> Result<()>;
 
+    #[allow(clippy::too_many_arguments)]
     // Similar to cudaMemcpy2D, though values are in elements and not in bytes.
     fn copy2d(
         &self,
@@ -107,6 +108,7 @@ pub trait BackendStorage: Sized {
         _d2: usize,
         _src_stride1: usize,
         _dst_stride1: usize,
+        _src_offset: usize,
         _dst_offset: usize,
     ) -> Result<()>;
 }