--- a/candle-core/src/cpu_backend.rs
+++ b/candle-core/src/cpu_backend.rs
@@ -1023,6 +1023,7 @@ impl<'a, I: IntDType> Map2 for IndexAdd<'a, I> {
     }
 }
 
+#[allow(clippy::too_many_arguments)]
 fn copy2d_<T: Copy>(
     src: &[T],
     dst: &mut [T],
@@ -1030,11 +1031,12 @@ fn copy2d_<T: Copy>(
     d2: usize,
     src_stride1: usize,
     dst_stride1: usize,
+    src_offset: usize,
     dst_offset: usize,
 ) {
     for i1 in 0..d1 {
         let dst_idx = i1 * dst_stride1 + dst_offset;
-        let src_idx = i1 * src_stride1;
+        let src_idx = i1 * src_stride1 + src_offset;
         let dst = &mut dst[dst_idx..dst_idx + d2];
         let src = &src[src_idx..src_idx + d2];
         dst.copy_from_slice(src)
@@ -2477,16 +2479,29 @@ impl BackendStorage for CpuStorage {
         d2: usize,
         src_s: usize,
         dst_s: usize,
+        src_o: usize,
         dst_o: usize,
     ) -> Result<()> {
         match (self, dst) {
-            (Self::U8(src), Self::U8(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
-            (Self::U32(src), Self::U32(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
-            (Self::I64(src), Self::I64(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
-            (Self::BF16(src), Self::BF16(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
-            (Self::F16(src), Self::F16(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
-            (Self::F32(src), Self::F32(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
-            (Self::F64(src), Self::F64(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, dst_o),
+            (Self::U8(src), Self::U8(dst)) => copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o),
+            (Self::U32(src), Self::U32(dst)) => {
+                copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o)
+            }
+            (Self::I64(src), Self::I64(dst)) => {
+                copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o)
+            }
+            (Self::BF16(src), Self::BF16(dst)) => {
+                copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o)
+            }
+            (Self::F16(src), Self::F16(dst)) => {
+                copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o)
+            }
+            (Self::F32(src), Self::F32(dst)) => {
+                copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o)
+            }
+            (Self::F64(src), Self::F64(dst)) => {
+                copy2d_(src, dst, d1, d2, src_s, dst_s, src_o, dst_o)
+            }
             (_, dst) => {
                 return Err(Error::DTypeMismatchBinaryOp {
                     lhs: self.dtype(),