--- a/candle-core/src/cuda_backend.rs
+++ b/candle-core/src/cuda_backend.rs
@@ -2157,22 +2157,18 @@ impl BackendStorage for CudaStorage {
         let dev = &self.device;
         let d1 = d1 as u32;
         let d2 = d2 as u32;
-        let src_s = src_s as u32;
-        let dst_s = dst_s as u32;
-        let (src, dst, elt_size_in_bytes) = match (&self.slice, &mut dst.slice) {
-            (S::U8(src), S::U8(dst)) => (*src.device_ptr(), *dst.slice(dst_o..).device_ptr(), 1),
-            (S::U32(src), S::U32(dst)) => (*src.device_ptr(), *dst.slice(dst_o..).device_ptr(), 4),
-            (S::BF16(src), S::BF16(dst)) => {
-                (*src.device_ptr(), *dst.slice(dst_o..).device_ptr(), 2)
-            }
-            (S::F16(src), S::F16(dst)) => (*src.device_ptr(), *dst.slice(dst_o..).device_ptr(), 2),
-            (S::F32(src), S::F32(dst)) => (*src.device_ptr(), *dst.slice(dst_o..).device_ptr(), 4),
-            (S::F64(src), S::F64(dst)) => (*src.device_ptr(), *dst.slice(dst_o..).device_ptr(), 8),
+        let (src, dst, elt_size) = match (&self.slice, &mut dst.slice) {
+            (S::U8(s), S::U8(d)) => (*s.device_ptr(), *d.slice(dst_o..).device_ptr(), 1),
+            (S::U32(s), S::U32(d)) => (*s.device_ptr(), *d.slice(dst_o..).device_ptr(), 4),
+            (S::BF16(s), S::BF16(d)) => (*s.device_ptr(), *d.slice(dst_o..).device_ptr(), 2),
+            (S::F16(s), S::F16(d)) => (*s.device_ptr(), *d.slice(dst_o..).device_ptr(), 2),
+            (S::F32(s), S::F32(d)) => (*s.device_ptr(), *d.slice(dst_o..).device_ptr(), 4),
+            (S::F64(s), S::F64(d)) => (*s.device_ptr(), *d.slice(dst_o..).device_ptr(), 8),
             _ => Err(CudaError::InternalError("dtype mismatch in copy2d"))?,
         };
         let func = dev.get_or_load_func("copy2d", kernels::FILL)?;
-        let cfg = LaunchConfig::for_num_elems(d1 * d2);
-        let params = (src, dst, d1, d2 * elt_size_in_bytes, src_s, dst_s);
+        let cfg = LaunchConfig::for_num_elems(d1);
+        let params = (src, dst, d1, d2 * elt_size, src_s as u32, dst_s as u32);
         // SAFETY: ffi.
         unsafe { func.launch(cfg, params) }.w()?;
         Ok(())