--- a/candle-core/src/storage.rs
+++ b/candle-core/src/storage.rs
@@ -701,4 +701,28 @@ impl Storage {
             .bt()),
         }
     }
+
+    pub(crate) fn copy2d(
+        &self,
+        dst: &mut Self,
+        d1: usize,
+        d2: usize,
+        src_s: usize,
+        dst_s: usize,
+        dst_o: usize,
+    ) -> Result<()> {
+        match (self, dst) {
+            (Self::Cpu(src), Self::Cpu(dst)) => src.copy2d(dst, d1, d2, src_s, dst_s, dst_o),
+            (Self::Cuda(src), Self::Cuda(dst)) => Ok(src.copy2d(dst, d1, d2, src_s, dst_s, dst_o)?),
+            (Self::Metal(src), Self::Metal(dst)) => {
+                Ok(src.copy2d(dst, d1, d2, src_s, dst_s, dst_o)?)
+            }
+            (lhs, rhs) => Err(Error::DeviceMismatchBinaryOp {
+                lhs: lhs.device().location(),
+                rhs: rhs.device().location(),
+                op: "copy2d",
+            }
+            .bt()),
+        }
+    }
 }